roles/k3s_server: ignore k3s-init.service start failure on non-first masters
Because we start all k3s-init processes at the same time and use
non-secure token, non-first masters will experience a transient failure
trying to fetch the CA bundle until the first master starts up and
starts serving it.

To hack around this, run non-first masters as Type=exec to ignore
service start failures and rely on systemd's Restart=. We are still
checking that the masters are up eventually, so this does not matter.