group_vars: rework external address logic and semantics
According to k3s, "external" address is, for example, a public address
that is DNATed to the node, and must never be equal to the internal
address.

Thus, having an "external interface" does not make sense because the
"external address" is something fully external to the node that it
cannot know by itself.

Due to the way Ansible variables work, we still default external IPs
to internal IPs but this is something that has to change.